requirements.txt
asgiref==3.7.2
Django==4.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.3.0
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
